Changed to use ES6 modules
