+++ examples/bar-colors.html
@@ -1,0 +1,85 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta http-equiv="Content-type" content="text/html; charset=utf-8">
+    <title>AmCharts</title>
+
+  </head>
+  <body id="" onload="">
+    <!-- Styles -->
+    <style>
+    .chart {
+      width: 100%;
+      height: 400px;
+    }
+    </style>
+
+    <!-- Resources -->
+    <script src="https://www.amcharts.com/lib/3/amcharts.js"></script>
+    <script src="https://www.amcharts.com/lib/3/serial.js"></script>
+    <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
+    <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
+    <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
+    <script src="../dist/cedar.js"></script>
+
+<!-- Chart code -->
+<script>
+var config1 = {
+  "type": "bar",
+  "datasets": [
+  {
+    "data": { "features":[
+      {"attributes": {"Type": "Elementary School", "Number_of_SUM": 200}},
+      {"attributes": {"Type": "Middle School", "Number_of_SUM": 100}},
+      {"attributes": {"Type": "High School", "Number_of_SUM": 50}}
+    ]},
+    "mappings": {
+      "category": {"field": "Type", "label": "Type"},
+      "series": [
+        {
+          "value": {"field": "Number_of_SUM"},
+          "label": "Additional Students"
+        }
+      ]
+    }
+  },
+  {
+    "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+    "query": {
+      "orderByFields": "Number_of_SUM DESC",
+      "groupByFieldsForStatistics": "Type",
+      "outStatistics": [
+        {
+          "statisticType": "sum",
+          "onStatisticField": "Number_of",
+          "outStatisticFieldName": "Number_of_SUM"
+        }
+      ]
+    },
+    "mappings": {
+      "category": {"field": "Type", "label": "Type"},
+      "series": [
+        {
+          "value": {"field": "Number_of_SUM"},
+          "label": "Number of Students"
+        }
+      ]
+    }
+  }],
+  "overrides": {
+    "categoryAxis": {
+      "labelRotation": -45
+    }
+  }
+
+}
+cedar("chartdiv1", config1);
+
+</script>
+
+<!-- HTML -->
+<a href="./">&lt; Back Home</a>
+<div id="chartdiv1" class="chart"></div>
+
+  </body>
+</html>
