+++ examples/custom.html
@@ -20,7 +20,7 @@
     <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
     <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
     <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
-    <script src="./cedar-amcharts.js"></script>
+    <script src="../dist/cedar.js"></script>
 
 <!-- Chart code -->
 <script>
@@ -166,7 +166,7 @@ var config1 = {
      }
   }
 }
-showChart("chartdiv1", config1);
+cedar("chartdiv1", config1);
 
 </script>
 
@@ -317,7 +317,7 @@ showChart("chartdiv1", config1);
          }
       }
     }
-    showChart("chartdiv1", config1);
+    cedar("chartdiv1", config1);
   </code>
 </pre>
   </body>
