+++ examples/injuries.html
@@ -20,7 +20,7 @@
     <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
     <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
     <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
-    <script src="./cedar-amcharts.js"></script>
+    <script src="../dist/cedar.js"></script>
 
 <!-- Chart code -->
 <a href="./">&lt; Back Home</a>
@@ -42,7 +42,7 @@ var config1 = {
   }
 }
 
-showChart("chartdiv1", config1);
+cedar("chartdiv1", config1);
 
 var config2 = {
       "type": "bar",
@@ -61,7 +61,7 @@ var config2 = {
       ]
   }
 }
-showChart("chartdiv2", config2);
+cedar("chartdiv2", config2);
 
 
 </script>
@@ -88,7 +88,7 @@ var config1 = {
   }
 }
 
-showChart("chartdiv1", config1);
+cedar("chartdiv1", config1);
 
 var config2 = {
       "type": "bar",
@@ -107,7 +107,7 @@ var config2 = {
       ]
   }
 }
-showChart("chartdiv2", config2);
+cedar("chartdiv2", config2);
 </pre>
 </body>
 </html>
