+++ examples/pie.html
@@ -10,22 +10,22 @@
     <style>
     .chart {
       width: 100%;
-      height: 400px;
+      height: 300px;
     }
     </style>
 
     <!-- Resources -->
     <script src="https://www.amcharts.com/lib/3/amcharts.js"></script>
-    <script src="https://www.amcharts.com/lib/3/serial.js"></script>
+    <script src="https://www.amcharts.com/lib/3/pie.js"></script>
     <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
     <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
     <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
-    <script src="./cedar-amcharts.js"></script>
+    <script src="../dist/cedar.js"></script>
 
 <!-- Chart code -->
 <script>
 var config1 = {
-  "type": "bar",
+  "type": "pie",
   "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
   "query": {
     "orderByFields": "Number_of_SUM DESC",
@@ -46,24 +46,21 @@ var config1 = {
         "label": "Number of Students"
       }
     ]
-  },
-  "overrides": {
-    "categoryAxis": {
-      "labelRotation": -45
-    }
   }
 }
-showChart("chartdiv1", config1);
+
+cedar("chartdiv1", config1);
 
 </script>
 
 <!-- HTML -->
 <a href="./">&lt; Back Home</a>
 <div id="chartdiv1" class="chart"></div>
+
 <pre>
   <code>
 var config1 = {
-  "type": "bar",
+  "type": "pie",
   "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
   "query": {
     "orderByFields": "Number_of_SUM DESC",
@@ -77,22 +74,16 @@ var config1 = {
     ]
   },
   "mappings": {
-    "category": {"field": Type, "label": Type},
+    "category": {"field": "Type", "label": "Type"},
     "series": [
       {
         "field": "Number_of_SUM",
         "label": "Number of Students"
       }
     ]
-  },
-  "overrides": {
-    "categoryAxis": {
-      "labelRotation": -45
-    }
   }
 }
-
-showChart("chartdiv1", config1);
+cedar("chartdiv1", config1);
   </code>
 </pre>
   </body>
