+++ examples/radar.html
@@ -20,7 +20,7 @@
     <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
     <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
     <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
-    <script src="./cedar-amcharts.js"></script>
+    <script src="../dist/cedar.js"></script>
 
 <!-- Chart code -->
 <script>
@@ -67,7 +67,7 @@ for(var p=0; p<time_periods.length; p++) {
     config1.datasets.push(s)
   }
 
-  showChart("chartdiv" + time_periods[p], config1);
+  cedar("chartdiv" + time_periods[p], config1);
 }
 </script>
