+++ examples/visionzerodc.html
@@ -22,7 +22,7 @@
     <script src="https://www.amcharts.com/lib/3/plugins/export/export.min.js"></script>
     <link rel="stylesheet" href="https://www.amcharts.com/lib/3/plugins/export/export.css" type="text/css" media="all" />
     <script src="https://www.amcharts.com/lib/3/themes/light.js"></script>
-    <script src="./cedar-amcharts.js"></script>
+    <script src="../dist/cedar.js"></script>
 
 <!-- Chart code -->
 <script>
@@ -54,7 +54,7 @@ var config1 = {
   }]
 }
 
-showChart("chartdiv1", config1);
+cedar("chartdiv1", config1);
 
 var time_periods = ["Hour", "Day", "Month"];
 var chart_types = ["radar", "bar", "line"];
@@ -86,7 +86,7 @@ for(var t=0; t<chart_types.length; t++) {
       }
     }]
   };
-  showChart("chartdiv" + time_periods[p] + chart_types[t], config1);
+  cedar("chartdiv" + time_periods[p] + chart_types[t], config1);
 
 
   }
