+++ index.js
@@ -1,0 +1,2 @@
+import cedar from './lib/cedar'
+export default cedar
