+++ lib/clone.js
@@ -1,0 +1,4 @@
+export default function(obj) {
+  console.log("clone", obj)
+  return JSON.parse(JSON.stringify(obj));
+}
