+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "cedar",
+  "version": "1.0.0",
+  "description": "Exploring visualization and analysis of DC VisionZero",
+  "main": "index.js",
+  "scripts": {
+    "start": "npm run build && concurrently \"watch 'npm run build' lib\" \"mocha -w\"",
+    "init": "mkdir dist",
+    "clean": "rm -rf dist",
+    "build": "rollup -c profiles/dev.js",
+    "build:release": "rollup -c profiles/dev.js && rollup -c profiles/prod.js",
+    "test": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/esridc/cedar-amcharts.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/esridc/cedar-amcharts/issues"
+  },
+  "homepage": "https://github.com/esridc/cedar-amcharts#readme",
+  "dependencies": {
+    "adlib": "0.0.5",
+    "rollup": "^0.41.6",
+    "rollup-plugin-buble": "^0.15.0",
+    "rollup-plugin-commonjs": "^8.0.2",
+    "rollup-plugin-node-resolve": "^3.0.0"
+  }
+}
